Add JoinLatticeTraits and start adapting the EdgeFunctions towards it
