+++ include/phasar/PhasarLLVM/DataFlowSolver/IfdsIde/SpecialSummaries.h
@@ -24,6 +24,7 @@
 #include <vector>
 
 #include "phasar/Config/Configuration.h"
+#include "phasar/PhasarLLVM/DataFlowSolver/IfdsIde/EdgeFunctionUtils.h"
 #include "phasar/PhasarLLVM/DataFlowSolver/IfdsIde/EdgeFunctions.h"
 #include "phasar/PhasarLLVM/DataFlowSolver/IfdsIde/FlowFunctions.h"
 #include "phasar/PhasarLLVM/Utils/BinaryDomain.h"
